requirements.txt
fastapi==0.116.1
uvicorn==0.35.0
xgboost==3.0.2
pandas==2.3.1
pydantic==2.11.7
python-dotenv==1.1.1
google-cloud-storage==3.2.0
scikit-learn==1.7.1
pytest==8.4.1
pytest-cov==6.2.1
httpx==0.28.1
locust==2.32.4